lib/requirements.txt
beautifulsoup4==4.11.1
discord.py==2.0.1
requests==2.28.0